+++ uspto/peds/tasks.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
 # (c) 2017 Andreas Motl <andreas@ip-tools.org>
 import celery
+from uspto.peds.client import UsptoPatentExaminationDataSystemClient
 from uspto.util.tasks import GenericDownloadTask
-from uspto.pbd.client import UsptoPairBulkDataClient
 
-class UsptoPairBulkDataDownloadTask(GenericDownloadTask):
-    name = 'uspto.pbd.tasks.UsptoPairBulkDataDownloadTask'
-    client_factory = UsptoPairBulkDataClient
+class UsptoPatentExaminationDataSystemDownloadTask(GenericDownloadTask):
+    name = 'uspto.peds.tasks.UsptoPatentExaminationDataSystemDownloadTask'
+    client_factory = UsptoPatentExaminationDataSystemClient
     pass
 
-@celery.shared_task(bind=True, base=UsptoPairBulkDataDownloadTask)
+@celery.shared_task(bind=True, base=UsptoPatentExaminationDataSystemDownloadTask)
 def download_task(self, query):
     """
     https://celery.readthedocs.io/en/latest/userguide/tasks.html#basics